# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/openings": {
            "get": {
                "description": "List all job opening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openings"
                ],
                "summary": "List Openings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/model.OpeningResponse"
							}
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
							"$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new job opening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openings"
                ],
                "summary": "Create opening",
                "parameters": [
                    {
						"description": "Opening data to create",
						"name": "opening",
                        "in": "body",
                        "required": true,
                        "schema": {
							"$ref": "#/definitions/model.OpeningRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
							"$ref": "#/definitions/model.OpeningResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
							"$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
					"404": {
						"description": "Not Found",
                        "schema": {
							"$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a new job opening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openings"
                ],
                "summary": "Delete opening",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Opening identification",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
							"$ref": "#/definitions/model.OpeningResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
							"$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
							"$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update a job opening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openings"
                ],
                "summary": "Update opening",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Opening Identification",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Opening data to Update",
                        "name": "opening",
                        "in": "body",
                        "required": true,
                        "schema": {
							"$ref": "#/definitions/model.OpeningRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
							"$ref": "#/definitions/model.OpeningResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
							"$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
							"$ref": "#/definitions/router.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
							"$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        },
		"/openings/:id": {
            "get": {
				"description": "Get first opening by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Openings"
                ],
				"summary": "Get Openings",
                "parameters": [
                    {
						"type": "string",
                        "description": "Opening identification",
                        "name": "id",
						"in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
							"$ref": "#/definitions/model.OpeningResponse"
                        }
                    },
					"500": {
						"description": "Internal Server Error",
                        "schema": {
							"$ref": "#/definitions/router.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
		"model.OpeningRequest": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "remote": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string"
                },
                "salary": {
                    "type": "integer"
                }
            }
        },
		"model.OpeningResponse": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "remote": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string"
                },
                "salary": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
				}
			}
		},
		"router.ErrorResponse": {
			"type": "object",
			"properties": {
				"errorCode": {
					"type": "string"
				},
				"message": {
					"type": "string"
                }
            }
        }
    }
}